I commented some classes that I haven't using It caused conflicts and also I change a table nameand and also I made changes in the user

+++ springFiles/src/main/java/com/example/orderswift/entity/OrderDetail.java
@@ -17,7 +17,7 @@ public class OrderDetail {
             name = "order_detail_id",
             updatable = false
     )
-    private Long orderDetailId;
+    private Integer orderDetailId;
 
     @ManyToOne
     @JoinColumn(name = "order_id")
